Split base classes (ScatterMaskTools) and image classes (MaskTools)
